Added uploading files to a user folder.
This task involved loading several files with checking for format compliance and checking for internal errors.

Also added is the ability to write the name of the folder in which the image will be saved. If the directory does not exist, a new one is created with the specified name.